+++ changelog.md
@@ -27,6 +27,7 @@
 - [actions/**] don't update integration_meta.sync_to_sap_succeeded if it is already true @maramalshen
 - [actions/create_invoice] include item.get_promotion in MEO_Serial & Promotion_Name @maramalshen
 - fix bug in is_matched in client & product @maramalshen
+- fix: update date formatting to use 'YYYYMMDD:000000' for UpdateAt field @maramalshen
 
 ### Removed